+++ src/client/viewmodels/index.ts
@@ -1,3 +1,5 @@
 export * from './activity.viewmodels';
+export * from './match.viewmodels';
 export * from './pair.viewmodels';
 export * from './profile.viewmodels';
+export * from './questionnaire.viewmodels';